Joins the local DATA_FLOW.md fixes/AI-data-connect work (previously on
a disconnected local history) with the real origin/feature/ai-integration
branch, using --allow-unrelated-histories so the remote's existing
commits are preserved rather than force-replaced.

Conflict resolution: kept the local (session) version for every file
except backend/app/schemas/canvas.py, where origin/feature/ai-integration
already had CanvasMetadata.width/height as int (a fix made independently
upstream) — kept that instead of reintroducing the local float version.
All other conflicting files were confirmed to be strictly-older remote
versions predating locally-added features (weak habits, target score,
stroke-order guide, practice-tab navigation) that this branch already
had before today's session started.

+++ frontend/lib/features/canvas_mode/models/stroke.dart
@@ -35,8 +35,8 @@ class Stroke {
 
 /// 캔버스 메타데이터 (SFR-003C Outputs 기준)
 class CanvasMetadata {
-  final double width;
-  final double height;
+  final int width;
+  final int height;
   final int strokeCount;
 
   const CanvasMetadata({
